requirements.txt
langchain_core==0.3.15
langchain_openai==0.2.5
python-dotenv==1.0.1
